Sharing the actual QHtml5CanvasPixmapData between local window and server was problematic: when drawing to a Pixmap, Qt first *copies* the QPixmap, so this would not have helped.  Instead, share the canvas handle, which is less hack-ish, anyway. Will need to ensure that this still works when we e.g. resize the local window (+ its pixmap) - I'm *guessing* Qt will take care of this, but don't know for sure, yet.
We now have the background widget colour being correctly drawn in the qtwebkit test server, so we're probably *finally* at the stage where the integration of our QHtml5CanvasPaintEngine with Qt is more-or-less 'done' (barring getting the window 'composition' to be more sophisticated, and hopefully sidestepping composition altogether when we can), and we can actually start working on it properly! Getting something drawn to the screen is far more involved than I would have expected!