Ok, after long deliberation, I've finally made a decision.
There are two approaches to the problem of HTML5 clips being additive (and thus, requiring us to restore() the context to the state before the clip was added if we want to set the clip to something different):

1) Take note of the actual clip, and undo it whenever we change pen or brush (or transform, etc) attributes, and re-clip to it when we come to do a drawing op;
2) When asked to clip, save all the pen, brush etc attributes; restore(); then re-set the pen, brush etc attributes again afterwards.

Both are approaches to stop the required restore() before we change clip from messing up all our pen and brush stuff, and both have disadvantages: 1) is complex, and also would   require us to somehow store clips that are more complex than just rectangles (as we will have to do when we implement clipping to QPainterPaths); 2) requires us to manually save and restore a growing list of properties everytime we clip.

On balance, I've decided to opt for 2) based a) on its comparative simplicity and b) on the hypothesis that there will be more drawing operations/ attribute changes than there will be clips in most apps.

Still extremely annoyed  that I have to resort to this, though!

The motivating example for all this is the newly added testChangingClipDoesntUndoSubsequentPenAndBrushChanges. As more and more pen and brush attribs are implemented, I'll have to write tests to ensure that they are not messed up clips.